test-task/app/requirements/requirements.txt
Django==2.1.7
gunicorn==19.9.0
psycopg2==2.7.5 --no-binary psycopg2
pytz
pillow==5.3.0
djangorestframework==3.8.2
django-rest-auth==0.9.3
djangorestframework-jwt==1.11.0 # PyJWT-1.6.4
coreapi==2.3.3
django-filter==2.0.0
pygments==2.3.1
markdown==3.0.1
django-crispy-forms==1.7.2
django-mptt==0.9.1
faker==1.0.4
